tkinter_app/requirements.txt
python-dotenv
peewee
psycopg2
paho-mqtt